update of oracle fixes from 2.1.x
git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@15799 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/plugin/oraclespatial/test/org/geotools/data/oracle/OracleDataStoreTest.java
@@ -20,10 +20,12 @@ import java.io.IOException;
 import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.sql.Types;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
@@ -53,6 +55,7 @@ import org.geotools.filter.Filter;
 import org.geotools.filter.FilterFactory;
 import org.geotools.filter.GeometryFilter;
 import org.geotools.filter.LikeFilter;
+import org.geotools.geometry.JTS.ReferencedEnvelope;
 
 import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.Envelope;
@@ -73,7 +76,7 @@ public class OracleDataStoreTest extends TestCase {
     private GeometryFactory jtsFactory = new GeometryFactory();
     private String schemaName;
     private OracleDataStore dstore;
-    
+    private Connection conn;
     /*
      * @see TestCase#setUp()
      */
@@ -84,11 +87,16 @@ public class OracleDataStoreTest extends TestCase {
         schemaName = properties.getProperty("schema");
         OracleConnectionFactory fact = new OracleConnectionFactory(properties.getProperty("host"), 
                 properties.getProperty("port"), properties.getProperty("instance"));
-        fact.setLogin(properties.getProperty("user"), properties.getProperty("passwd"));
-        cPool = fact.getConnectionPool();
-        Connection conn = cPool.getConnection();
+        fact.setLogin(properties.getProperty("user"), properties.getProperty("passwd"));        
+        try {
+        	cPool = fact.getConnectionPool();                    
+            conn = cPool.getConnection();
+        } catch( Throwable t ){
+        	System.out.println("Warning: could not connect, configure "+getClass().getResource("test.properties"));
+        	return;
+        }
         System.out.println(conn.getTypeMap());
-        
+        reset();
         dstore = new OracleDataStore(cPool, properties.getProperty("schema"), new HashMap());
         dstore.setFIDMapper("ORA_TEST_POINTS", new TypedFIDMapper(new MaxIncFIDMapper("ORA_TEST_POINTS", "ID", Types.INTEGER), "ORA_TEST_POINTS"));
         dstore.setFIDMapper("ORA_TEST_LINES", new TypedFIDMapper(new MaxIncFIDMapper("ORA_TEST_LINES", "ID", Types.INTEGER), "ORA_TEST_LINES"));
@@ -98,63 +106,118 @@ public class OracleDataStoreTest extends TestCase {
     /*
      * @see TestCase#tearDown()
      */
-    protected void tearDown() throws Exception {
-        Connection conn = cPool.getConnection();
-        Statement st = conn.createStatement();
-        st.executeUpdate("DELETE FROM ORA_TEST_POINTS");
-        st.executeUpdate("INSERT INTO ora_test_points VALUES ('point 1',10,1," +
-                "MDSYS.SDO_GEOMETRY(2001,NULL,NULL," +
-                "MDSYS.SDO_ELEM_INFO_ARRAY(1,1,1)," +
-                "MDSYS.SDO_ORDINATE_ARRAY(10,10)))");
-        st.executeUpdate("INSERT INTO ora_test_points VALUES ('point 2',20,2," +
-                        "MDSYS.SDO_GEOMETRY(2001,NULL,NULL," +
-                        "MDSYS.SDO_ELEM_INFO_ARRAY(1,1,1)," +
-                        "MDSYS.SDO_ORDINATE_ARRAY(20,10)))");
-        st.executeUpdate("INSERT INTO ora_test_points VALUES ('point 3',30,3," +
-                        "MDSYS.SDO_GEOMETRY(2001,NULL,NULL," +
-                        "MDSYS.SDO_ELEM_INFO_ARRAY(1,1,1)," +
-                        "MDSYS.SDO_ORDINATE_ARRAY(20,30)))");
-        st.executeUpdate("INSERT INTO ora_test_points VALUES ('point 4',40,4," +
-                        "MDSYS.SDO_GEOMETRY(2001,NULL,NULL," +
-                        "MDSYS.SDO_ELEM_INFO_ARRAY(1,1,1)," +
-                        "MDSYS.SDO_ORDINATE_ARRAY(30,10)))");
-        st.executeUpdate("INSERT INTO ora_test_points VALUES ('point 5',50,5," +
-                        "MDSYS.SDO_GEOMETRY(2001,NULL,NULL," +
-                        "MDSYS.SDO_ELEM_INFO_ARRAY(1,1,1)," +
-                        "MDSYS.SDO_ORDINATE_ARRAY(-20,10)))");
+    protected void tearDown() throws Exception {    	
+    	//reset();
         ConnectionPoolManager manager = ConnectionPoolManager.getInstance();
-        manager.closeAll();
+        manager.closeAll();    	
+    }
+    static boolean first  = false;
+    
+    protected void reset() throws Exception {
+    	if( conn == null ) return;
+    	Statement st = conn.createStatement();
+    	
+    	if( !first ){
+    		try {
+    			st.execute("DROP TABLE ora_test_points");
+    			st.executeUpdate("DELETE FROM user_sdo_geom_metadata WHERE TABLE_NAME='ORA_TEST_POINTS'");    	    	    			
+    		}
+    		catch( SQLException noPrevRun){}
+    		first = true;
+    	}
+    	try {
+			if( st.executeQuery("SELECT TABLE_NAME FROM USER_TABLES WHERE TABLE_NAME = 'ORA_TEST_POINTS'").next()){
+				try {
+		    		st.executeUpdate("DELETE FROM ORA_TEST_POINTS");
+		    	}
+		    	catch (SQLException fine){
+		    		// must be the first time?
+		    		fine.printStackTrace();
+		    	}    					
+			}
+			else {
+				st.execute( "CREATE TABLE ORA_TEST_POINTS ("+
+						    "       NAME VARCHAR2(15),"+
+						    "       INTVAL NUMBER(12,0),"+"" +
+						    "       ID NUMBER(3,0) PRIMARY KEY,"+
+						    "       SHAPE MDSYS.SDO_GEOMETRY )");
+		    	st.execute("INSERT INTO user_sdo_geom_metadata (TABLE_NAME, COLUMN_NAME, DIMINFO, SRID )"+
+					    "VALUES('ORA_TEST_POINTS','SHAPE',"+
+					        "MDSYS.SDO_DIM_ARRAY(MDSYS.SDO_DIM_ELEMENT('X',0,-20,0.5),MDSYS.SDO_DIM_ELEMENT('Y',0,10,0.5)),"+
+					    "NULL)");
+//		    	 st.execute("DROP INDEX GEOTOOLS.ORA_TEST_POINTS_SHAPE_IDX");		    	
+		    	st.execute("CREATE INDEX GEOTOOLS.ORA_TEST_POINTS_SHAPE_IDX "+
+		    				"ON GEOTOOLS.ORA_TEST_POINTS (SHAPE) INDEXTYPE IS " +
+		    				"MDSYS.SPATIAL_INDEX PARAMETERS (' SDO_INDX_DIMS=2 LAYER_GTYPE=\"COLLECTION\"') ");		    			    	
+			}
+    	}
+    	catch (SQLException fine){
+    		fine.printStackTrace();
+    	}  	
+    	//                   + (20,30)
+    	//         
+    	//  +(20,10)         +(10,10)   + (20,10)   +(30,10)
+    	//        
+    	st.execute("INSERT INTO ORA_TEST_POINTS VALUES ('point 1',10,1," +
+                "MDSYS.SDO_GEOMETRY(2001,NULL," +
+                "MDSYS.SDO_POINT_TYPE(10.0, 10.0, NULL),"+
+                "NULL,NULL))");
+    	st.execute("INSERT INTO ORA_TEST_POINTS VALUES ('point 2',20,2," +
+                        "MDSYS.SDO_GEOMETRY(2001,NULL," +
+                        "MDSYS.SDO_POINT_TYPE(20.0, 10.0, NULL),"+
+                        "NULL,NULL))");
+    	st.execute("INSERT INTO ORA_TEST_POINTS VALUES ('point 3',30,3," +
+                        "MDSYS.SDO_GEOMETRY(2001,NULL," +
+                        "SDO_POINT_TYPE(20.0, 30.0, NULL),"+
+                        "NULL,NULL))");
+    	st.execute("INSERT INTO ORA_TEST_POINTS VALUES ('point 4',40,4," +
+                        "MDSYS.SDO_GEOMETRY(2001,NULL," +
+                        "SDO_POINT_TYPE(30.0, 10.0, NULL),"+
+                        "NULL,NULL))");
+    	st.execute("INSERT INTO ORA_TEST_POINTS VALUES ('point 5',50,5," +
+                        "MDSYS.SDO_GEOMETRY(2001,NULL," +
+                        "SDO_POINT_TYPE(-20.0, 10.0, NULL),"+
+                        "NULL,NULL))");
+
+    	
     }
 
     public void testGetFeatureTypes() throws IOException {
+    	if( conn == null ) return;
+    	
         String[] fts = dstore.getTypeNames();
-        System.out.println(Arrays.asList(fts));
-        assertEquals(3, fts.length);
+        List list = Arrays.asList( fts );
+        
+        System.out.println( list );
+        assertTrue( list.contains("ORA_TEST_POINTS"));        
     }
 
     public void testGetSchema() throws Exception {
+    	if( conn == null ) return;    	
             FeatureType ft = dstore.getSchema("ORA_TEST_POINTS");
             assertNotNull(ft);
             System.out.println(ft);
     }
 
     public void testGetFeatureReader() throws Exception {
+    	if( conn == null ) return;    	
             FeatureType ft = dstore.getSchema("ORA_TEST_POINTS");
             Query q = new DefaultQuery( "ORA_TEST_POINTS" );
             FeatureReader fr = dstore.getFeatureReader( q, Transaction.AUTO_COMMIT);
+            assertEquals( ft, fr.getFeatureType() );
             int count = 0;
 
             while (fr.hasNext()) {
                 fr.next();
                 count++;
             }
-
             assertEquals(5, count);
 
             fr.close();
     }
 
     public void testGetFeatureWriter() throws Exception {
+    	if( conn == null ) return;    	
         FeatureWriter writer = dstore.getFeatureWriter("ORA_TEST_POINTS", Filter.NONE, Transaction.AUTO_COMMIT);
         assertNotNull(writer);
 
@@ -178,21 +241,22 @@ public class OracleDataStoreTest extends TestCase {
         reader.close();
     }
     
-//    public void testMaxFeatures() throws Exception {        
-//        DefaultQuery query = new DefaultQuery();
-//        query.setTypeName("ORA_TEST_POINTS");
-//        query.setMaxFeatures(3);
-//        DataStore ds = new OracleDataStore(cPool, "SEAN");
-//        FeatureSource fs = ds.getFeatureSource("ORA_TEST_POINTS");
-//        FeatureResults fr = fs.getFeatures(query);
-//        assertEquals(3, fr.getCount());
-//    }
+    public void testMaxFeatures() throws Exception {
+    	if( conn == null ) return;    	
+        DefaultQuery query = new DefaultQuery();
+        query.setTypeName("ORA_TEST_POINTS");
+        query.setMaxFeatures(3);
+        FeatureSource fs = dstore.getFeatureSource("ORA_TEST_POINTS");        
+        FeatureResults fr = fs.getFeatures(query);
+        assertEquals(3, fr.getCount());
+    }
 
     public void testLikeGetFeatures() throws Exception {
+    	if( conn == null ) return;    	
         LikeFilter likeFilter = filterFactory.createLikeFilter();
         Expression pattern = filterFactory.createLiteralExpression("*");
         Expression attr = filterFactory.createAttributeExpression(null, "NAME");
-        likeFilter.setPattern(pattern, "*", "?", "\\");
+        likeFilter.setPattern(pattern, "*", "?", "\\"); // '*' --> '%' 
         likeFilter.setValue(attr);
         
         FeatureSource fs = dstore.getFeatureSource("ORA_TEST_POINTS");
@@ -205,7 +269,8 @@ public class OracleDataStoreTest extends TestCase {
         assertEquals(1, fr.getCount());
     }
     
-    public void testAttributeFilter() throws Exception {        
+    public void testAttributeFilter() throws Exception {
+    	if( conn == null ) return;    	
         CompareFilter attributeEquality = filterFactory.createCompareFilter(AbstractFilter.COMPARE_EQUALS);
         Expression attribute = filterFactory.createAttributeExpression(dstore.getSchema("ORA_TEST_POINTS"), "NAME");
         Expression literal = filterFactory.createLiteralExpression("point 1");
@@ -223,6 +288,7 @@ public class OracleDataStoreTest extends TestCase {
     }
     
     public void testBBoxFilter() throws Exception {
+    	if( conn == null ) return;    	
         GeometryFilter filter = filterFactory.createGeometryFilter(AbstractFilter.GEOMETRY_BBOX);
         Expression right = filterFactory.createBBoxExpression(new Envelope(-180, 180, -90, 90));
         Expression left = filterFactory.createAttributeExpression(dstore.getSchema("ORA_TEST_POINTS"), "SHAPE");
@@ -231,15 +297,20 @@ public class OracleDataStoreTest extends TestCase {
         
         FeatureSource fs = dstore.getFeatureSource("ORA_TEST_POINTS");
         FeatureResults fr = fs.getFeatures(filter);        
-        assertEquals(5, fr.getCount());
+        assertEquals(5, fr.getCount()); // we pass this!
         
-        right = filterFactory.createBBoxExpression(new Envelope(15, 35, 0, 15));
+    	//                   + (20,30)
+    	//                            +----------------------+
+    	//  +(20,10)         +(10,10) | + (20,10)   +(30,10) |
+    	//                            +----------------------+
+        right = filterFactory.createBBoxExpression(new Envelope(15, 35, 0, 15));        
         filter.addRightGeometry(right);
         fr = fs.getFeatures(filter);
-        assertEquals(2, fr.getCount());
+        assertEquals(2, fr.getCount()); // we have 4!
     }
     
     public void testPointGeometryConversion() throws Exception {
+    	if( conn == null ) return;    	
         CompareFilter filter = filterFactory.createCompareFilter(AbstractFilter.COMPARE_EQUALS);
         Expression left = filterFactory.createAttributeExpression(dstore.getSchema("ORA_TEST_POINTS"), "NAME");
         Expression right = filterFactory.createLiteralExpression("point 1");
@@ -260,6 +331,7 @@ public class OracleDataStoreTest extends TestCase {
     }
     
     public void testAddFeatures() throws Exception {
+    	if( conn == null ) return;    	
         Map fidGen = new HashMap();
         fidGen.put("ORA_TEST_POINTS", JDBCDataStoreConfig.FID_GEN_MANUAL_INC);
         JDBCDataStoreConfig config = JDBCDataStoreConfig.createWithSchemaNameAndFIDGenMap(schemaName, fidGen);
@@ -286,6 +358,7 @@ public class OracleDataStoreTest extends TestCase {
     }
     
     public void testRemoveFeatures() throws Exception {
+    	if( conn == null ) return;    	
         FeatureStore fs = (FeatureStore) dstore.getFeatureSource("ORA_TEST_POINTS");
         fs.removeFeatures(Filter.NONE);
         FeatureResults fr = fs.getFeatures();
@@ -293,6 +366,7 @@ public class OracleDataStoreTest extends TestCase {
     }
     
     public void testPropertySelect() throws Exception {
+    	if( conn == null ) return;    	
         DefaultQuery q = new DefaultQuery("ORA_TEST_POINTS",Filter.NONE);
         q.setPropertyNames(new String[]{"NAME"});
         FeatureReader fr = dstore.getFeatureReader(q, Transaction.AUTO_COMMIT);
@@ -300,5 +374,13 @@ public class OracleDataStoreTest extends TestCase {
         FeatureType ft = f.getFeatureType();
         assertEquals(1, ft.getAttributeCount());
         assertEquals("NAME", ft.getAttributeType(0).getName());        
+    }
+    public void testBounds(){
+    	if( conn == null ) return;    	    	
+    	Envelope extent = dstore.getEnvelope("ORA_TEST_POINTS");
+    	assertNotNull( extent );
+    	assertTrue( extent instanceof ReferencedEnvelope );
+    	ReferencedEnvelope envelope = (ReferencedEnvelope) extent;
+    	assertFalse( envelope.isNull() );
     }
 }
